+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### v 2.7 (WIP)
-- new: file mamager
+- new: file manager
+- new: intro screen
+- new: about screen
 
 ### v 2.6
 - fixed: f-droid buid error